better arrayfield pattern, add input components
