+++ .github/dependabot.yml
@@ -3,5 +3,5 @@ updates:
     - package-ecosystem: "npm"
       directory: "/"
       schedule:
-          interval: "daily"
+          interval: "weekly"
       open-pull-requests-limit: 10
