+++ public/sitemap.xml
@@ -9,74 +9,74 @@
     <url
         ><loc>https://umwics.vercel.app/about</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/contact</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/coop</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/carousels</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/companies</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/events</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/members</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/resources</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs/users</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/docs</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/events/[id]</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/events</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/members/[id]</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/members</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/mentors</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/outreach</loc>
 
-        <lastmod>2020-10-12</lastmod> </url
+        <lastmod>2020-10-25</lastmod> </url
     ><url
         ><loc>https://umwics.vercel.app/resources</loc>
 
-        <lastmod>2020-10-12</lastmod>
+        <lastmod>2020-10-25</lastmod>
     </url></urlset
 >
