+++ src/components/inputs/DateInput.tsx
@@ -1,0 +1,9 @@
+import { Field, FieldAttributes } from "formik";
+import { KeyboardDatePicker } from "formik-material-ui-pickers";
+import React from "react";
+
+const DateInput: React.FC<FieldAttributes<any>> = (props: FieldAttributes<any>) => {
+    return <Field {...props} component={KeyboardDatePicker} />;
+};
+
+export default DateInput;
