+++ src/pages/admin/members/index.tsx
@@ -3,8 +3,8 @@ import { GetStaticProps, NextPage } from "next";
 import { useSnackbar } from "notistack";
 import { useState } from "react";
 import { DropResult, ResponderProvided } from "react-beautiful-dnd";
-import AddMemberDialog from "../../../components/admin/AddMemberDialog";
-import MemberList from "../../../components/admin/MemberList";
+import AddMemberDialog from "../../../components/admin/forms/AddMemberDialog";
+import MemberList from "../../../components/admin/views/MemberList";
 import { useConfirm } from "../../../components/ConfirmProvider";
 import AdminLayout from "../../../components/layouts/AdminLayout";
 import { hasPermission, Member, User } from "../../../interfaces";
